Created Group model backend for grouping students and teachers wrote tests for group api as well as behavior such as getting the accumilated points of it's members
fixed async tests bug with the JSReporter plugin for Jasmine
wrote a workaround for validating/publishing/saving challanges by having the publish command always revalidate, this is because observing the change to the challenge properties cased the model to invaldidate when ever it's saved so it would never be valid. may change once I figure out how.

Applied access control based on roles in the front end tho students can still access the view of the create page if they have a link (they will not be able to create from the backend check tho)

fixed regex bug in firefox
made info toast when not logged in only pop up different messages dependign on challange completion