Teachers can now assign quests to students or groups implimented a few API ports in in quest in order to acomidate for that most notibally PUT 'api/quests/:id/assign' which takes data {users:[ids], groups:[ids]} to assign wrote tests for new API ports implimented CRUD for userQuest
Changed realtonship name in Quest and User to userQuests previously users and quests respectively

considers allowing students to view and take on quests on their own, which would require quests to be publishable and somethought may be put into that.

for now need to finish up how quests are vied by students and how they contribute to the user score
need to list quests assigned to student
need to show student progress in quest for teacher
need to do notification for all these events (can wait on making them only for group).
deploy all these changes